Rename file back to HelloWorld. It's not a good name, but makes the PR more useful
